plz work

+++ BadStringAutomaton.cpp
@@ -25,7 +25,7 @@ void BadStringAutomaton::S1(const std::string& input) {
         S2(input);
     }
     //Sends error if reach the end of file.
-    else if(input.size() == index){
+    else if((int)input.size() == index){
         return;
     }
     else{
